+++ schema/productionSeries.ts
