+++ schema/video.ts
